Move javax.usb and native stuff into their own packages.
The libusb wrapper of usb4java is now considered to be the central
main part of usb4java and the usb4java project itself only contains
the Java code of this library. The javax.usb stuff is moved into
the usb4java-javax project and the native JNI code is moved into the
libusb4java project. Binary files are no longer present as Java
resources. Instead they are (hopefully) automatically generated and
published as separate JAR files (And Maven Dependencies) so they can
be easily and selectively included into a project.